requirements.txt
click==4.0
oauthlib==0.7.2
Pillow==2.8.1
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
rq==0.5.2
six==1.9.0
tweepy==3.3.0
